requirements.txt
# Chess LLM Benchmark Dependencies

# Core
python-chess>=1.10.0
requests>=2.31.0
pyyaml>=6.0.1

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
filelock>=3.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
